Issue #1: resolve Test Failure
There were three issues to be resolved. The get request was requesting the lastTradedPrice for AAPL and the test was expecting the symbol to be FB. This was resolved by changing the get request to request for FB instead of AAPL.

Additionally the expected price was different from what was being return for the price in the mapping-lastTradedPrice.json. This was changed to match the returned request.

Finally, the mapping-lastTradedPrice.json was return a status of 404 despite it being a successful request. This was changed from a status 404 to status 200.