Issue #2: Add new Endpoint for Historical Prices
Create a new method in IexRestController getHistoricalPrice.

Inputs:
Symbols (required): the stock ticker symbols that historical price data will be queried for
Date (not required, default = today): the date at which to get the historical prices for
Ranges (not required, default = 1 month): the range from the date to get the historical prices for

IexRestController will serve as the endpoint for users of FUSE.

The method will call iexService which will have a new method getHistoricalPrices, this method will utilise the iexClient to fetch the data from the api.

The IexService getHistoricalPrices method will contain logic to call the correct method from the IexCloudClient depending on the amount of arguments passed in.

The IexCloudClient will be a new interface which will connect and make calls  to the Iex Cloud api. It will contain multiple methods which will be called by IexService.

The return value will be a list of IexHistoricalPrice objects which will contain the specified data for each day.