Updated Historical Price endpoint based on feedback
Added annotation validation to IexrestController that will respond with a bad request if symbol is empty. Updated IexService logic to remove the handling for this condition as it will no longer reach it.

Additionally, changed testGetHistoricalPriceEmpty to reflect these changes.

Removed unused imports in IexClient and IexCloudClient.

Updated IexRestControllerTest to reduce the path for MockMvcRequestBuilders.